terraform/tf/requirements.txt
tfparse
matplotlib
python-hcl2
networkx